requirements.txt
python-dotenv==0.19.1
pandas==1.3.4
sqlalchemy==1.4.25
